[update] 서비스 코드 수정

+++ helloDaeyeo/src/main/java/com/daeyeo/helloDaeyeo/service/RentalLogService.java
@@ -30,7 +30,7 @@ public class RentalLogService {
      * @param userId
      */
 
-    public void insertRentalLog(RentalLogDto rentalLogDto, int rentalId , String userId){
+    public void insertRentalLog(RentalLogDto rentalLogDto, long rentalId , String userId){
         Optional<RentalObject> rentalObject = rentalObjectRepository.findById(rentalId);
         Optional<Member> member = memberRepository.findById(userId);
         RentalLog rentalLog = new RentalLog(rentalLogDto);
